4 components removed - 3 zooanimals  and 1 select
