--- client/src/components/NocturnalZooComponent.css
@@ -1,3 +1,0 @@
-.dark {
-    background-color: mediumpurple;
-}
