--- client/src/components/ZooSelect.js
@@ -1,7 +1,0 @@
-const ZooSelect = ()=> {
-    return(
-        <h3>This is a select component</h3>
-    )
-}
-
-export default ZooSelect;
